UserControllerを作成し、PostControllerに記述していたユーザーに関する処理をお引越し。メソッド名とルーティングの名前を修正。バリーデーション時にはじかれたユーザー名とメアドを表示させるように条件分岐を追加
